* develop: (146 commits)
  Global Event Contribution Popup: Development - First implementation.
  Update human village, water cave and witch forest scene
  Added 2 new pets
  dragon fat WIP
  dragon fat WIP
  Update water caves, witch forest and fog.
  Update fog
  update fog and fortress scenes
  adjust the wood
  Update lighting and fog.
  Fixes
  adjust layout
  Fixes: prefab names
  Fixes
  Update light map for all scenes.
  User Profile: Development - Unified all currencies logic.
  Deleted some prefabs not used
  mines final test
  water area under castle almost finished. Some changed needed in the caves where there are goblins
  First Candidate for Central Area!
  ...


Former-commit-id: 7f63c327a40e9118852bcac8d79ef88561c58806

+++ Assets/Game/Model/Eggs/EggManager.cs
@@ -86,7 +86,7 @@ public class EggManager : UbiBCN.SingletonMonoBehaviour<EggManager> {
 	}
 
 	public static int goldenEggFragments {
-		get { return instance.m_user.goldenEggFragments; }
+		get { return (int)instance.m_user.goldenEggFragments; }
 	}
 
 	public static bool goldenEggCompleted {
